Visualization now resizing properly
This works well on desktops, but on Android it causes audio to
crackle. Might need to test moving the visualization code into
a web worker.

Also needs the ability to toggle the visualization on/off to save
battery on mobile.

Finally also needs to wait for Streamlit version 1.32.3 which
fixes iframe background color on dark mode.